refactor(utils): move get_os_info() to an utils.os module
This code was not in utils.py, but its tests were in test_utils.py, so
move them all.

Also move client related tests to test_client.py.
