+++ ggshield/utils/os.py
@@ -1,0 +1,50 @@
+import logging
+import platform
+import sys
+from functools import lru_cache
+from pathlib import Path
+from typing import Tuple
+
+
+logger = logging.getLogger(__name__)
+
+
+@lru_cache(None)
+def get_os_info() -> Tuple[str, str]:
+    """
+    Returns a tuple of [OS, version]
+
+    OS is always in lowercase
+    """
+    if sys.platform.lower() == "linux":
+        return parse_os_release(Path("/etc/os-release"))
+    else:
+        return platform.system().lower(), platform.version()
+
+
+def parse_os_release(os_release_path: Path) -> Tuple[str, str]:
+    """
+    Extract and return Linux's OS-name and OS-Version
+    If extraction fails, we return ('linux', 'unknown')
+    """
+    error_tuple = "linux", "unknown"
+
+    try:
+        with open(os_release_path) as f:
+            lines = f.readlines()
+
+        # Build a dictionary from the os-release file contents
+        data_dict = {}
+        for line in lines:
+            if "=" in line:
+                key, value = line.split("=")
+                key, value = key.strip(), value.strip()
+                data_dict[key] = value.strip('"')
+
+        if "ID" not in data_dict:
+            return error_tuple
+
+        return data_dict["ID"], data_dict.get("VERSION_ID", "unknown")
+    except Exception as exc:
+        logger.warning(f"Failed to read Linux OS name and version: {exc}")
+        return error_tuple
